Improved variables begin regex.
To avoid matching strings like `Something <= OtherThing` where the
`<` of the comparison operator was mistaken for a generic, require
there to be at least one storage type and at least one identifier.

Dollar are valid identifiers, added that in a few places it was
missing.

Also simplifies some regexes. E.g. `\w` matches `a-zA-Z0-9_` so
some redundant chars could be removed.