Fix regression of field values being incorrectly highlighted (#161)
### Requirements

* Filling out the template is required. Any pull request that does not include enough information to be reviewed in a timely manner may be closed at the maintainers' discretion.
* All new code requires tests to ensure against regressions

### Description of the Change

This PR fixes the regression reported in #158. This happened because our variable definition now ends with `=`, and the rest is supposed to be matched by the code, which does happen correctly, but it breaks in certain scenarios when using class fields, e.g.
```java
class A {
  private String b = a + "a()" + a() + "" + "a();" + "" + a() + abc + b() + "b();";
  private String c = "a / a();";
  private int d[] = a + "a();" + func();

  int abcd() {
    return 1;
  }
}
```

### Before
<img width="691" alt="before" src="https://user-images.githubusercontent.com/7788766/45589511-114e3000-b927-11e8-98c9-94b52dc37f8f.png">

### After
<img width="691" alt="after" src="https://user-images.githubusercontent.com/7788766/45589512-13b08a00-b927-11e8-8083-3eb0fd051069.png">


All of the `b`, `c`, and `d` would break a subsequent highlighting and scope of method `abcd`.

The fix is fairly straightforward and might seem as a hack. We define `#class-fields-and-methods` scope instead of `#methods` and first match anything like this `... = [...];`, where `[...]` is covered by `#code`, and only then we match methods. This seems to work quite well.

I also fixed comments in methods, while I was there. Added tests for both methods and class fields.

### Alternate Designs

I considered reverting #146 at some point, but then we would still have to solve `;` issue in try with resources, and, to be fair, the change in that PR was more or less reasonable.

I also tried updating pattern for `#methods` scope trying to restrict it, so that it does not show up in the expression part. It turned out to be quite difficult, mostly because of the return type and parameter types being potentially anything, which complicated the task. For example, I tried writing something like this: 
```cson
'methods':
    'begin': '(?!new)(?=[\\w<].*\\s+)(?=[\\w\\s\\[\\]<>,\\.\\?/\\*]+\\()'
```
But it did not work in certain cases.

### Benefits

Fixes the regression introduced by #146.

### Possible Drawbacks

Could potentially miss coverage of some other case.

### Applicable Issues

Closes #158