fix enum with overrides (#175)
### Requirements

* Filling out the template is required. Any pull request that does not include enough information to be reviewed in a timely manner may be closed at the maintainers' discretion.
* All new code requires tests to ensure against regressions

### Description of the Change

This PR extends support for enums and adds a pattern to handle enum's constant fields with method overrides. It is allowed to have such fields mixed with the previous two types that we have added.

The fix is fairly simple - adding another pattern to handle the case of enum field starting with `{` and ending with `}` similar to inner class. For that reason, I used `include: #class-body` instead of `include: #code`.

### Alternate Designs

None were considered, since it is fairly simple fix.

### Benefits

Fixes enum highlighting for enums with override methods.

### Possible Drawbacks

None. It only affects changes for enums.

### Applicable Issues

Closes #173