add extends and implements to enum (#190)
### Requirements

* Filling out the template is required. Any pull request that does not include enough information to be reviewed in a timely manner may be closed at the maintainers' discretion.
* All new code requires tests to ensure against regressions

### Description of the Change

This PR adds `extends` and `implements` to `enums` scope. Currently we have enum scope separate from class and interface, because of the complexity of enum's representations, though we did not migrate `extends` and `implements` to the enum. This work fixes the issue by copying those blocks of code from class scope and adding them to enum scope with some minor changes.

### Alternate Designs

None were considered.

### Benefits

Fixes a regression for enum's `extends` and `implements`.

### Possible Drawbacks

This patch introduces a bit of code duplication between class and enum, mostly because of the separation of those scopes.

### Applicable Issues

Closes #185