Fixes order of storage modifiers, so @interface is scoped properly inside a class
### Requirements

* Filling out the template is required. Any pull request that does not include enough information to be reviewed in a timely manner may be closed at the maintainers' discretion.
* All new code requires tests to ensure against regressions

### Description of the Change

This PR fixes the issue when `@interface` defined within a class with a storage modifier would not be scoped properly. We change the order of `storage-modifiers` scope to make sure it gets applied before variables scope.

### Alternate Designs

None were considered.

### Benefits

Fixes scope and highlighting of `@interface` within a class.

### Possible Drawbacks

I had to update one of the tests for class fields. Since we apply storage modifiers before variables, we now correctly parse `int` in the test example.

### Applicable Issues

Closes #203