Highlight comments in module (#246)
Signed-off-by: Yan Zhang <yanzh@microsoft.com>

### Requirements

* Filling out the template is required. Any pull request that does not include enough information to be reviewed in a timely manner may be closed at the maintainers' discretion.
* All new code requires tests to ensure against regressions

### Description of the Change

<!--

We must be able to understand the design of your change from this description. If we can't get a good idea of what the code will be doing from the description here, the pull request may be closed at the maintainers' discretion. Keep in mind that the maintainer reviewing this PR may not be familiar with or have worked with the code here recently, so please walk us through the concepts.

-->

To enable comments and javadocs in `module`

### Alternate Designs
unknown
<!-- Explain what other alternates were considered and why the proposed version was selected -->

### Benefits
To fix part of #245 
<!-- What benefits will be realized by the code change? -->

### Possible Drawbacks
unknown
<!-- What are the possible side-effects or negative impacts of the code change? -->

### Applicable Issues

#245 
<!-- Enter any applicable Issues here -->